chore: clean repo structure and remove Eclipse metadata from tracking
Move .git to project root, remove .metadata/.classpath/.project/.settings
from tracking, remove .env.development and impressoras.config from index.
Code changes: web BFF tenant-aware routes, API auth updates, desktop DAO fixes.

+++ naviera-api/src/main/java/com/naviera/api/controller/CadastrosController.java
@@ -120,7 +120,7 @@ public class CadastrosController {
         return ResponseEntity.ok(writeService.criarUsuario(TenantUtils.getEmpresaId(auth), dados));
     }
     @PutMapping("/usuarios/{id}")
-    public ResponseEntity<?> atualizarUsuario(@PathVariable Long id, @RequestBody Map<String, Object> dados, Authentication auth) {
+    public ResponseEntity<?> atualizarUsuario(@PathVariable Integer id, @RequestBody Map<String, Object> dados, Authentication auth) {
         return ResponseEntity.ok(writeService.atualizarUsuario(TenantUtils.getEmpresaId(auth), id, dados));
     }
