+++ naviera-api/src/main/java/com/naviera/api/dto/UsuarioDTO.java
@@ -1,3 +1,3 @@
 package com.naviera.api.dto;
 
-public record UsuarioDTO(Long id, String nome, String login, String email, String funcao, String permissoes) {}
+public record UsuarioDTO(Integer id, String nome, String login, String email, String funcao, String permissoes) {}
